refactor into indep and subsamplemessenger
